fs: Allow fine-grained control of folio sizes
Some filesystems want to be able to limit the maximum size of folios,
and some want to be able to ensure that folios are at least a certain
size.  Add mapping_set_folio_orders() to allow this level of control
(although it is not yet honoured).  Rename FS_THP_SUPPORT to
FS_MULTI_PAGE_FOLIO to reflect the new naming.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>