fs: Convert page_mkwrite_check_truncate to folio_mkwrite_check_truncate
If a filesystem uses multi-page folios then its ->page_mkwrite() method
may be called with a tail page.  If it is, the filesystem should treat
the entire folio as being made dirty, so it must iterate over every
extent covered by the folio in case it needs to be COWed.

Change both callers to pass the folio of the page.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>