mm/compaction: fix misbehaviors of fast_find_migrateblock()
In the fast_find_migrateblock(), It iterate freelist to find proper
pageblock.  But there are some misbehaviors.

First, if the page we found is equal to cc->migrate_pfn, it is considered
that we didn't found suitable pageblock.  Second, if the loop was
terminated because order is less than PAGE_ALLOC_COSTLY_ORDER, it could be
considered that we found suitable one.  Third, if the skip bit is set on
the page block and go continue, it doesn't check the nr_scanned.  Fourth,
if the page block's skip bit is set, it check that page block is the last
of list.  But it is unnecessary.

Link: https://lkml.kernel.org/r/20210128130411.6125-1-vvghjk1234@gmail.com
Fixes: 70b44595eafe9 ("mm, compaction: use free lists to quickly locate a migration source")
Signed-off-by: Wonhyuk Yang <vvghjk1234@gmail.com>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Mel Gorman <mgorman@techsingularity.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>