mm/swap: Convert put_pages_list to use folios
Pages on the LRU list cannot be tail pages as ->compound_head is in
a union with one of the words of the list_head, and they cannot be
ZONE_DEVICE pages as ->pgmap is in a union with the same word.  So add
lru_to_folio() and use folio_put() in put_pages_list().  Saves 86 bytes
of text.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>